Moved out AudioSamplePool
